@@ -1,10 +1,9 @@
 module github.com/docker/compose-cli
 
-go 1.16
+go 1.19
 
 require (
-	github.com/AlecAivazis/survey/v2 v2.2.3
-	github.com/Azure/azure-pipeline-go v0.2.2 // indirect
+	github.com/AlecAivazis/survey/v2 v2.3.6
 	github.com/Azure/azure-sdk-for-go v48.2.0+incompatible
 	github.com/Azure/azure-storage-file-go v0.8.0
 	github.com/Azure/go-autorest/autorest v0.11.12
