@@ -21,7 +21,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
@@ -31,23 +30,34 @@ import (
 	"github.com/docker/cli/cli/config/configfile"
 	"github.com/docker/docker/api/types"
 	dockerclient "github.com/docker/docker/client"
+	"github.com/hashicorp/go-version"
 	"github.com/spf13/pflag"
 )
 
-// getBuildMetadata returns build metadata for this command
-func getBuildMetadata(cliSource string, command string, args []string) string {
+// BuildMetadata returns build metadata for this command
+func BuildMetadata(cliSource, cliVersion, command string, args []string) string {
 	var cli, builder string
 	dockercfg := config.LoadDefaultConfigFile(io.Discard)
 	if alias, ok := dockercfg.Aliases["builder"]; ok {
+		if alias != "buildx" {
+			return cliSource
+		}
 		command = alias
 	}
 	if command == "build" {
-		cli = "docker"
-		builder = "buildkit"
-		if enabled, _ := isBuildKitEnabled(); !enabled {
-			builder = "legacy"
+		buildkitEnabled, _ := isBuildKitEnabled()
+		if buildkitEnabled && isBuildxDefault(cliVersion) {
+			command = "buildx"
+			args = append([]string{"build"}, args...)
+		} else {
+			cli = "docker"
+			builder = "buildkit"
+			if !buildkitEnabled {
+				builder = "legacy"
+			}
 		}
-	} else if command == "buildx" {
+	}
+	if command == "buildx" {
 		cli = "buildx"
 		builder = buildxDriver(dockercfg, args)
 	}
@@ -104,7 +114,7 @@ func buildxDriver(dockercfg *configfile.ConfigFile, buildArgs []string) string {
 		//   "Name": "builder",
 		//   "Global": false
 		// }
-		rawCurrent, err := ioutil.ReadFile(fileCurrent)
+		rawCurrent, err := os.ReadFile(fileCurrent)
 		if err != nil {
 			return driver
 		}
@@ -156,7 +166,7 @@ func buildxDriver(dockercfg *configfile.ConfigFile, buildArgs []string) string {
 	//   ],
 	//   "Dynamic": false
 	// }
-	rawBuilder, err := ioutil.ReadFile(fileBuilder)
+	rawBuilder, err := os.ReadFile(fileBuilder)
 	if err != nil {
 		return driver
 	}
@@ -174,16 +184,33 @@ func buildxDriver(dockercfg *configfile.ConfigFile, buildArgs []string) string {
 // buildxBuilder returns the builder being used in the build command
 func buildxBuilder(buildArgs []string) string {
 	var builder string
-	fset := pflag.NewFlagSet("buildx", pflag.ContinueOnError)
-	fset.String("builder", "", "")
-	_ = fset.ParseAll(buildArgs, func(flag *pflag.Flag, value string) error {
-		if flag.Name == "builder" {
-			builder = value
-		}
-		return nil
-	})
+	flags := pflag.NewFlagSet("buildx", pflag.ContinueOnError)
+	flags.Usage = func() {}
+	flags.StringVar(&builder, "builder", "", "")
+	_ = flags.Parse(buildArgs)
 	if len(builder) == 0 {
 		builder = os.Getenv("BUILDX_BUILDER")
 	}
 	return builder
+}
+
+// isBuildxDefault returns true if buildx by default is used
+// through "docker build" command which is already an alias to
+// "docker buildx build" in docker cli.
+// more info: https://github.com/docker/cli/pull/3314
+func isBuildxDefault(cliVersion string) bool {
+	if cliVersion == "" {
+		// empty means DWARF symbol table is stripped from cli binary
+		// which is the case with docker cli < 22.06
+		return false
+	}
+	verCurrent, err := version.NewVersion(cliVersion)
+	if err != nil {
+		return false
+	}
+	// 21.0.0 is an arbitrary version number because next major is not
+	// intended to be 21 but 22 and buildx by default will never be part
+	// of a 20 release version anyway.
+	verBuildxDefault, _ := version.NewVersion("21.0.0")
+	return verCurrent.GreaterThanOrEqual(verBuildxDefault)
 }
