@@ -19,26 +19,60 @@ package metrics
 import (
 	"os"
 	"strings"
+	"sync"
+	"time"
+
+	"github.com/docker/compose/v2/pkg/utils"
 
 	"github.com/docker/compose-cli/cli/metrics/metadata"
-	"github.com/docker/compose/v2/pkg/utils"
 )
 
-// Track sends the tracking analytics to Docker Desktop
-func Track(context string, args []string, status string) {
+func (c *client) Track(cmd CmdResult) {
 	if isInvokedAsCliBackend() {
 		return
 	}
-	command := GetCommand(args)
-	if command != "" {
-		c := NewClient()
-		c.Send(Command{
-			Command: command,
-			Context: context,
-			Source:  metadata.Get(CLISource, args),
-			Status:  status,
-		})
+
+	var wg sync.WaitGroup
+	usageCmd := NewCommandUsage(cmd)
+	if usageCmd != nil {
+		usageCmd.Source = c.getMetadata(CLISource, cmd.Args)
+		wg.Add(1)
+		go func() {
+			c.reporter.Heartbeat(*usageCmd)
+			wg.Done()
+		}()
 	}
+
+	eventCmd := NewDockerCLIEvent(cmd)
+	if eventCmd != nil {
+		wg.Add(1)
+		go func() {
+			c.reporter.Event(*eventCmd)
+			wg.Done()
+		}()
+	}
+
+	done := make(chan struct{})
+	go func() {
+		defer close(done)
+		wg.Wait()
+	}()
+
+	select {
+	case <-done:
+	case <-time.After(Timeout):
+	}
+}
+
+func (c *client) getMetadata(cliSource string, args []string) string {
+	if len(args) == 0 {
+		return cliSource
+	}
+	switch args[0] {
+	case "build", "buildx":
+		cliSource = metadata.BuildMetadata(cliSource, c.cliversion.f(), args[0], args[1:])
+	}
+	return cliSource
 }
 
 func isInvokedAsCliBackend() bool {
@@ -89,4 +123,17 @@ func GetCommand(args []string) string {
 		}
 	}
 	return result
+}
+
+func NewCommandUsage(cmd CmdResult) *CommandUsage {
+	command := GetCommand(cmd.Args)
+	if command == "" {
+		return nil
+	}
+
+	return &CommandUsage{
+		Command: command,
+		Context: cmd.ContextType,
+		Status:  cmd.Status,
+	}
 }
