@@ -18,15 +18,18 @@ package mobycli
 
 import (
 	"context"
+	"debug/buildinfo"
 	"fmt"
 	"os"
 	"os/exec"
 	"os/signal"
 	"path/filepath"
 	"regexp"
+	"runtime"
+	"strings"
+	"time"
 
-	"github.com/docker/compose/v2/pkg/compose"
-	"github.com/docker/compose/v2/pkg/utils"
+	"github.com/google/shlex"
 	"github.com/spf13/cobra"
 
 	apicontext "github.com/docker/compose-cli/api/context"
@@ -38,7 +41,13 @@ import (
 var delegatedContextTypes = []string{store.DefaultContextType}
 
 // ComDockerCli name of the classic cli binary
-const ComDockerCli = "com.docker.cli"
+var ComDockerCli = "com.docker.cli"
+
+func init() {
+	if runtime.GOOS == "windows" {
+		ComDockerCli += ".exe"
+	}
+}
 
 // ExecIfDefaultCtxType delegates to com.docker.cli if on moby context
 func ExecIfDefaultCtxType(ctx context.Context, root *cobra.Command) {
@@ -63,45 +72,65 @@ func mustDelegateToMoby(ctxType string) bool {
 }
 
 // Exec delegates to com.docker.cli if on moby context
-func Exec(root *cobra.Command) {
+func Exec(_ *cobra.Command) {
+	metricsClient := metrics.NewDefaultClient()
+	metricsClient.WithCliVersionFunc(func() string {
+		return CliVersion()
+	})
+	start := time.Now().UTC()
 	childExit := make(chan bool)
 	err := RunDocker(childExit, os.Args[1:]...)
 	childExit <- true
+	duration := time.Since(start)
 	if err != nil {
 		if exiterr, ok := err.(*exec.ExitError); ok {
 			exitCode := exiterr.ExitCode()
-			metrics.Track(store.DefaultContextType, os.Args[1:], compose.ByExitCode(exitCode).MetricsStatus)
+			metricsClient.Track(
+				metrics.CmdResult{
+					ContextType: store.DefaultContextType,
+					Args:        os.Args[1:],
+					Status:      metrics.FailureCategoryFromExitCode(exitCode).MetricsStatus,
+					ExitCode:    exitCode,
+					Start:       start,
+					Duration:    duration,
+				},
+			)
 			os.Exit(exitCode)
 		}
-		metrics.Track(store.DefaultContextType, os.Args[1:], compose.FailureStatus)
+		metricsClient.Track(
+			metrics.CmdResult{
+				ContextType: store.DefaultContextType,
+				Args:        os.Args[1:],
+				Status:      metrics.FailureStatus,
+				Start:       start,
+				Duration:    duration,
+			},
+		)
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 	}
 	commandArgs := os.Args[1:]
 	command := metrics.GetCommand(commandArgs)
-	if command == "build" && !metrics.HasQuietFlag(commandArgs) {
-		utils.DisplayScanSuggestMsg()
-	}
 	if command == "login" && !metrics.HasQuietFlag(commandArgs) {
 		displayPATSuggestMsg(commandArgs)
 	}
-	metrics.Track(store.DefaultContextType, os.Args[1:], compose.SuccessStatus)
+	metricsClient.Track(
+		metrics.CmdResult{
+			ContextType: store.DefaultContextType,
+			Args:        os.Args[1:],
+			Status:      metrics.SuccessStatus,
+			ExitCode:    0,
+			Start:       start,
+			Duration:    duration,
+		},
+	)
 
 	os.Exit(0)
 }
 
 // RunDocker runs a docker command, and forward signals to the shellout command (stops listening to signals when an event is sent to childExit)
 func RunDocker(childExit chan bool, args ...string) error {
-	execBinary, err := resolvepath.LookPath(ComDockerCli)
-	if err != nil {
-		execBinary = findBinary(ComDockerCli)
-		if execBinary == "" {
-			fmt.Fprintln(os.Stderr, err)
-			fmt.Fprintln(os.Stderr, "Current PATH : "+os.Getenv("PATH"))
-			os.Exit(1)
-		}
-	}
-	cmd := exec.Command(execBinary, args...)
+	cmd := exec.Command(comDockerCli(), args...)
 	cmd.Stdin = os.Stdin
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
@@ -131,6 +160,25 @@ func RunDocker(childExit chan bool, args ...string) error {
 	return cmd.Run()
 }
 
+func comDockerCli() string {
+	if v := os.Getenv("DOCKER_COM_DOCKER_CLI"); v != "" {
+		return v
+	}
+
+	execBinary := findBinary(ComDockerCli)
+	if execBinary == "" {
+		var err error
+		execBinary, err = resolvepath.LookPath(ComDockerCli)
+		if err != nil {
+			fmt.Fprintln(os.Stderr, err)
+			fmt.Fprintln(os.Stderr, "Current PATH : "+os.Getenv("PATH"))
+			os.Exit(1)
+		}
+	}
+
+	return execBinary
+}
+
 func findBinary(filename string) string {
 	currentBinaryPath, err := os.Executable()
 	if err != nil {
@@ -149,7 +197,7 @@ func findBinary(filename string) string {
 
 // IsDefaultContextCommand checks if the command exists in the classic cli (issues a shellout --help)
 func IsDefaultContextCommand(dockerCommand string) bool {
-	cmd := exec.Command(ComDockerCli, dockerCommand, "--help")
+	cmd := exec.Command(comDockerCli(), dockerCommand, "--help")
 	b, e := cmd.CombinedOutput()
 	if e != nil {
 		fmt.Println(e)
@@ -157,12 +205,41 @@ func IsDefaultContextCommand(dockerCommand string) bool {
 	return regexp.MustCompile("Usage:\\s*docker\\s*" + dockerCommand).Match(b)
 }
 
+// CliVersion returns the docker cli version
+func CliVersion() string {
+	info, err := buildinfo.ReadFile(ComDockerCli)
+	if err != nil {
+		return ""
+	}
+	for _, s := range info.Settings {
+		if s.Key != "-ldflags" {
+			continue
+		}
+		args, err := shlex.Split(s.Value)
+		if err != nil {
+			return ""
+		}
+		for _, a := range args {
+			// https://github.com/docker/cli/blob/f1615facb1ca44e4336ab20e621315fc2cfb845a/scripts/build/.variables#L77
+			if !strings.HasPrefix(a, "github.com/docker/cli/cli/version.Version") {
+				continue
+			}
+			parts := strings.Split(a, "=")
+			if len(parts) != 2 {
+				return ""
+			}
+			return parts[1]
+		}
+	}
+	return ""
+}
+
 // ExecSilent executes a command and do redirect output to stdOut, return output
 func ExecSilent(ctx context.Context, args ...string) ([]byte, error) {
 	if len(args) == 0 {
 		args = os.Args[1:]
 	}
-	cmd := exec.CommandContext(ctx, ComDockerCli, args...)
+	cmd := exec.CommandContext(ctx, comDockerCli(), args...)
 	cmd.Stderr = os.Stderr
 	return cmd.Output()
 }
