@@ -60,7 +60,7 @@ func TestAllMethodsHaveCorrespondingCliCommand(t *testing.T) {
 
 func TestTrackSuccess(t *testing.T) {
 	var mockMetrics = &mockMetricsClient{}
-	mockMetrics.On("Send", metrics.Command{Command: "ps", Context: "aci", Status: "success", Source: "api"}).Return()
+	mockMetrics.On("SendUsage", metrics.CommandUsage{Command: "ps", Context: "aci", Status: "success", Source: "api"}).Return()
 	newClient := client.NewClient("aci", noopService{})
 	interceptor := metricsServerInterceptor(mockMetrics)
 
@@ -122,6 +122,14 @@ type mockMetricsClient struct {
 	mock.Mock
 }
 
-func (s *mockMetricsClient) Send(command metrics.Command) {
+func (s *mockMetricsClient) WithCliVersionFunc(f func() string) {
+	s.Called(f)
+}
+
+func (s *mockMetricsClient) SendUsage(command metrics.CommandUsage) {
 	s.Called(command)
+}
+
+func (s *mockMetricsClient) Track(cmd metrics.CmdResult) {
+	s.Called(cmd)
 }
