@@ -19,9 +19,10 @@ package ecs
 import (
 	"context"
 
-	"github.com/docker/compose-cli/utils"
 	"github.com/docker/compose/v2/pkg/api"
 	"github.com/docker/compose/v2/pkg/progress"
+
+	"github.com/docker/compose-cli/utils"
 )
 
 func (b *ecsAPIService) Down(ctx context.Context, projectName string, options api.DownOptions) error {
