@@ -1,3 +1,4 @@
+//go:build kube
 // +build kube
 
 /*
@@ -22,10 +23,11 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/docker/compose-cli/api/context/store"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/client-go/tools/clientcmd/api"
+
+	"github.com/docker/compose-cli/api/context/store"
 )
 
 // ListAvailableKubeConfigContexts list kube contexts
