@@ -21,10 +21,11 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/docker/compose-cli/api/config"
 	"github.com/docker/compose/v2/pkg/api"
 	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
+
+	"github.com/docker/compose-cli/api/config"
 )
 
 // CheckUnsupported checks if a flag was used when it shouldn't and adds an error in case
